order captures first
Bench: 9776977
